+++ src/main/kotlin/com/xaquare/xquarebackoffice/infrastructure/excel/service/GetUserInfo.kt
@@ -1,23 +1,29 @@
 package com.xaquare.xquarebackoffice.infrastructure.excel.service
 
-import org.springframework.stereotype.Service
+import com.xaquare.xquarebackoffice.domain.entity.User
+import com.xaquare.xquarebackoffice.domain.persistence.UserPersistenceAdapter
 import org.apache.poi.xssf.usermodel.XSSFWorkbook
-import org.apache.poi.ss.usermodel.FillPatternType
-import org.apache.poi.ss.usermodel.IndexedColors
+import org.springframework.stereotype.Service
 import javax.servlet.http.HttpServletResponse
 import org.apache.poi.ss.usermodel.BorderStyle
 import org.apache.poi.ss.usermodel.Cell
 import org.apache.poi.ss.usermodel.CellStyle
+import org.apache.poi.ss.usermodel.FillPatternType
+import org.apache.poi.ss.usermodel.IndexedColors
 import org.apache.poi.ss.usermodel.Row
 import org.apache.poi.ss.usermodel.Sheet
 import org.apache.poi.ss.usermodel.Workbook
+import org.springframework.transaction.annotation.Transactional
 
 @Service
-class CreateExcelSheetService {
+class GetUserInfo(
+    private val userPersistenceAdapter: UserPersistenceAdapter
+) {
 
+    @Transactional(readOnly = true)
     fun execute(response: HttpServletResponse) {
         val workbook: Workbook = XSSFWorkbook()
-        val sheet: Sheet = workbook.createSheet("xquare").apply {
+        val sheet: Sheet = workbook.createSheet("xquare_userInfo").apply {
             defaultColumnWidth = 30
         }
 
@@ -33,7 +39,7 @@ class CreateExcelSheetService {
             )
         }
 
-        val headerNames = arrayOf("이름", "아이디", "비밀번호", "학년", "반", "번호", "프로필 사진")
+        val headerNames = arrayOf("이름", "아이디", "비밀번호", "학년", "반", "번호", "프로필사진")
 
         val headerRow: Row = sheet.createRow(0)
         headerNames.forEachIndexed { i, header ->
@@ -48,23 +54,25 @@ class CreateExcelSheetService {
             setBorderStyle(BorderStyle.THIN)
         }
 
-        val bodyData = arrayOf(
-            arrayOf("예시) 홍길동", "예시) gildong", "예시) Gildong1!", "예시) 1", "예시) 1", "예시) 1", "예시) http:// "),
-        )
+        val userList: List<User> = userPersistenceAdapter.findAll()
 
-        bodyData.forEachIndexed { i, bodyRowData ->
-            val bodyRow: Row = sheet.createRow(i + 1)
-            bodyRowData.forEachIndexed { j, data ->
-                val bodyCell: Cell = bodyRow.createCell(j).apply {
-                    setCellValue(data)
-                    cellStyle = bodyCellStyle
-                }
+        userList.forEachIndexed { index, user ->
+            val bodyRow: Row = sheet.createRow(index + 1)
+            bodyRow.createCell(0).setCellValue(user.name)
+            bodyRow.createCell(1).setCellValue(user.accountId)
+            bodyRow.createCell(2).setCellValue(user.password)
+            bodyRow.createCell(3).setCellValue(user.grade.toDouble())
+            bodyRow.createCell(4).setCellValue(user.classNum.toDouble())
+            bodyRow.createCell(5).setCellValue(user.num.toDouble())
+            bodyRow.createCell(6).setCellValue(user.profile)
+            bodyRow.forEach { cell ->
+                cell.cellStyle = bodyCellStyle
             }
         }
 
         // File
-        val fileName = "xquare.spreadsheetml_download"
-        response.contentType = "application/xquare.spreadsheetml.sheet"
+        val fileName = "xquare_userInfo.spreadsheetml_download"
+        response.contentType = "application/xquare_userInfo.spreadsheetml.sheet"
         response.setHeader("Content-Disposition", "attachment;filename=$fileName.xlsx")
         val servletOutputStream = response.outputStream
